+++ packages/client/src/aa/create.ts
@@ -4,7 +4,7 @@ import type { AAProvider } from "./types";
 import { createAlchemyAAState } from "./alchemy/create";
 import { createPimlicoAAState } from "./pimlico/create";
 import type { AAOwner } from "./owner";
-import type { AAMode, AAState, WalletCall } from "./types";
+import type { AAMode, AAState, AAWalletCall } from "./types";
 
 export type { AAOwner } from "./owner";
 
@@ -13,7 +13,7 @@ export interface CreateAAStateOptions {
   chain: Chain;
   owner: AAOwner;
   rpcUrl: string;
-  callList: WalletCall[];
+  callList: AAWalletCall[];
   mode?: AAMode;
   apiKey?: string;
   gasPolicyId?: string;
