+++ packages/client/src/aa/types.ts
@@ -12,10 +12,10 @@ export type AASponsorship = "disabled" | "optional" | "required";
 // Call Types
 // ---------------------------------------------------------------------------
 
-export type WalletCall = {
-  to: string;
-  value: string;
-  data?: string;
+export type AAWalletCall = {
+  to: Hex;
+  value: bigint;
+  data?: Hex;
   chainId: number;
 };
 
@@ -23,12 +23,6 @@ export type WalletAtomicCapability = {
   atomic?: {
     status?: string;
   };
-};
-
-export type WalletPrimitiveCall = {
-  to: Hex;
-  value: bigint;
-  data?: Hex;
 };
 
 // ---------------------------------------------------------------------------
@@ -68,13 +62,16 @@ export interface AAResolvedConfig {
 // Provider Abstractions
 // ---------------------------------------------------------------------------
 
+/** The subset of AAWalletCall passed to smart account send methods (chainId already resolved). */
+export type AACallPayload = Omit<AAWalletCall, "chainId">;
+
 export interface SmartAccount {
   provider: string;
   mode: string;
   AAAddress?: Hex;
   delegationAddress?: Hex;
-  sendTransaction: (call: WalletPrimitiveCall) => Promise<{ transactionHash: string }>;
-  sendBatchTransaction: (calls: WalletPrimitiveCall[]) => Promise<{ transactionHash: string }>;
+  sendTransaction: (call: AACallPayload) => Promise<{ transactionHash: string }>;
+  sendBatchTransaction: (calls: AACallPayload[]) => Promise<{ transactionHash: string }>;
 }
 
 export interface AAState<TAccount extends SmartAccount = SmartAccount> {
@@ -99,7 +96,7 @@ export interface ExecutionResult {
 }
 
 export interface AtomicBatchArgs {
-  calls: WalletPrimitiveCall[];
+  calls: AACallPayload[];
   capabilities?: {
     atomic?: {
       required?: boolean;
@@ -110,7 +107,7 @@ export interface AtomicBatchArgs {
 export interface ExecuteWalletCallsParams<
   TAccount extends SmartAccount = SmartAccount,
 > {
-  callList: WalletCall[];
+  callList: AAWalletCall[];
   currentChainId: number;
   capabilities: Record<string, WalletAtomicCapability> | undefined;
   localPrivateKey: `0x${string}` | null;
@@ -148,7 +145,7 @@ export const SPONSORSHIP_MODES = new Set<AASponsorship>([
 
 export function getAAChainConfig(
   config: AAConfig,
-  calls: WalletCall[],
+  calls: AAWalletCall[],
   chainsById: Record<number, Chain>,
 ): AAChainConfig | null {
   if (!config.enabled || calls.length === 0) {
@@ -218,14 +215,6 @@ export function getWalletExecutorReady(
 // Execution
 // ---------------------------------------------------------------------------
 
-export function mapCall(call: WalletCall): WalletPrimitiveCall {
-  return {
-    to: call.to as Hex,
-    value: BigInt(call.value),
-    data: call.data ? (call.data as Hex) : undefined,
-  };
-}
-
 export const DEFAULT_AA_CONFIG: AAConfig = {
   enabled: true,
   provider: "alchemy",
@@ -281,230 +270,3 @@ export const DISABLED_PROVIDER_STATE: AAState = {
   error: null,
 };
 
-export async function executeWalletCalls(
-  params: ExecuteWalletCallsParams,
-): Promise<ExecutionResult> {
-  const {
-    callList,
-    currentChainId,
-    capabilities,
-    localPrivateKey,
-    providerState,
-    sendCallsSyncAsync,
-    sendTransactionAsync,
-    switchChainAsync,
-    chainsById,
-    getPreferredRpcUrl,
-  } = params;
-
-  if (providerState.resolved && providerState.account) {
-    return executeViaAA(callList, providerState);
-  }
-
-  if (providerState.resolved && providerState.error && !providerState.resolved.fallbackToEoa) {
-    throw providerState.error;
-  }
-
-  return executeViaEoa({
-    callList,
-    currentChainId,
-    capabilities,
-    localPrivateKey,
-    sendCallsSyncAsync,
-    sendTransactionAsync,
-    switchChainAsync,
-    chainsById,
-    getPreferredRpcUrl,
-  });
-}
-
-// ---------------------------------------------------------------------------
-// Internal Execution Paths
-// ---------------------------------------------------------------------------
-
-async function executeViaAA(
-  callList: WalletCall[],
-  providerState: AAState,
-): Promise<ExecutionResult> {
-  const account = providerState.account;
-  const resolved = providerState.resolved;
-
-  if (!account || !resolved) {
-    throw providerState.error ?? new Error("smart_account_unavailable");
-  }
-
-  const callsPayload = callList.map(mapCall);
-  const receipt =
-    callList.length > 1
-      ? await account.sendBatchTransaction(callsPayload)
-      : await account.sendTransaction(callsPayload[0]);
-  const txHash = receipt.transactionHash;
-  const providerPrefix = account.provider.toLowerCase();
-
-  // For 7702, the SDK may not provide the delegation address (or provide the
-  // EOA which is already filtered out by adaptSmartAccount).  Fall back to
-  // reading the authorization list from the on-chain transaction.
-  let delegationAddress: Hex | undefined =
-    account.mode === "7702" ? account.delegationAddress : undefined;
-
-  if (account.mode === "7702" && !delegationAddress) {
-    delegationAddress = await resolve7702Delegation(txHash, callList);
-  }
-
-  return {
-    txHash,
-    txHashes: [txHash],
-    executionKind: `${providerPrefix}_${account.mode}`,
-    batched: callList.length > 1,
-    sponsored: resolved.sponsorship !== "disabled",
-    AAAddress: account.AAAddress,
-    delegationAddress,
-  };
-}
-
-/**
- * Best-effort extraction of the 7702 delegation target from the on-chain
- * transaction's authorization list.  Returns `undefined` on any failure so
- * the caller can safely fall through.
- */
-async function resolve7702Delegation(
-  txHash: string,
-  callList: WalletCall[],
-): Promise<Hex | undefined> {
-  try {
-    const { createPublicClient, http } = await import("viem");
-    const chainId = callList[0]?.chainId;
-    if (!chainId) return undefined;
-
-    const { mainnet, polygon, arbitrum, optimism, base } = await import("viem/chains");
-    const knownChains: Record<number, Parameters<typeof createPublicClient>[0]["chain"]> = {
-      1: mainnet, 137: polygon, 42161: arbitrum, 10: optimism, 8453: base,
-    };
-    const chain = knownChains[chainId];
-    if (!chain) return undefined;
-
-    const client = createPublicClient({ chain, transport: http() });
-    const tx = await client.getTransaction({ hash: txHash as Hex });
-    const authList = (
-      tx as unknown as {
-        authorizationList?: Array<{ address?: Hex; contractAddress?: Hex }>;
-      }
-    ).authorizationList;
-    const target = authList?.[0]?.address ?? authList?.[0]?.contractAddress;
-    if (target) {
-      return target;
-    }
-  } catch {
-    // Best-effort — don't fail the whole execution.
-  }
-  return undefined;
-}
-
-async function executeViaEoa({
-  callList,
-  currentChainId,
-  capabilities,
-  localPrivateKey,
-  sendCallsSyncAsync,
-  sendTransactionAsync,
-  switchChainAsync,
-  chainsById,
-  getPreferredRpcUrl,
-}: Omit<ExecuteWalletCallsParams, "providerState">): Promise<ExecutionResult> {
-  const { createPublicClient, createWalletClient, http } = await import("viem");
-  const { privateKeyToAccount } = await import("viem/accounts");
-
-  const hashes: string[] = [];
-
-  if (localPrivateKey) {
-    for (const call of callList) {
-      const chain = chainsById[call.chainId];
-      if (!chain) {
-        throw new Error(`Unsupported chain ${call.chainId}`);
-      }
-      const rpcUrl = getPreferredRpcUrl(chain);
-      if (!rpcUrl) {
-        throw new Error(`No RPC for chain ${call.chainId}`);
-      }
-
-      const account = privateKeyToAccount(localPrivateKey);
-      const walletClient = createWalletClient({
-        account,
-        chain,
-        transport: http(rpcUrl),
-      });
-      const hash = await walletClient.sendTransaction({
-        account,
-        to: call.to as Hex,
-        value: BigInt(call.value),
-        data: call.data ? (call.data as Hex) : undefined,
-      });
-      const publicClient = createPublicClient({
-        chain,
-        transport: http(rpcUrl),
-      });
-      await publicClient.waitForTransactionReceipt({ hash });
-      hashes.push(hash);
-    }
-
-    return {
-      txHash: hashes[hashes.length - 1],
-      txHashes: hashes,
-      executionKind: "eoa",
-      batched: hashes.length > 1,
-      sponsored: false,
-    };
-  }
-
-  const chainIds = Array.from(new Set(callList.map((call) => call.chainId)));
-  if (chainIds.length > 1) {
-    throw new Error("mixed_chain_bundle_not_supported");
-  }
-
-  const chainId = chainIds[0];
-  if (currentChainId !== chainId) {
-    await switchChainAsync({ chainId });
-  }
-
-  const chainCaps = capabilities?.[`eip155:${chainId}`];
-  const atomicStatus = chainCaps?.atomic?.status;
-  const canUseSendCalls = atomicStatus === "supported" || atomicStatus === "ready";
-
-  if (canUseSendCalls) {
-    const batchResult = await sendCallsSyncAsync({
-      calls: callList.map(mapCall),
-      capabilities: {
-        atomic: {
-          required: true,
-        },
-      },
-    });
-
-    const receipts =
-      (batchResult as { receipts?: Array<{ transactionHash?: string }> }).receipts ??
-      [];
-    for (const receipt of receipts) {
-      if (receipt.transactionHash) {
-        hashes.push(receipt.transactionHash);
-      }
-    }
-  } else {
-    for (const call of callList) {
-      const hash = await sendTransactionAsync({
-        chainId: call.chainId,
-        to: call.to as Hex,
-        value: BigInt(call.value),
-        data: call.data ? (call.data as Hex) : undefined,
-      });
-      hashes.push(hash);
-    }
-  }
-
-  return {
-    txHash: hashes[hashes.length - 1],
-    txHashes: hashes,
-    executionKind: "eoa",
-    batched: hashes.length > 1,
-    sponsored: false,
-  };
-}
